api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
gunicorn==22.0.0
pydantic==2.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
reportlab==4.2.0
qrcode==7.4.2
Pillow==10.3.0
requests==2.32.3
google-cloud-tasks==2.16.4
google-cloud-logging==3.10.0

# Testing dependencies
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
pytest-mock==3.14.0
factory-boy==3.3.0